requirements.txt
pyspark==3.5.5
delta-spark==3.1.0
requests==2.31.0
kafka-python-ng